+++ src/app/app.filter.pipe.ts
@@ -6,12 +6,13 @@ import { escapeRegExp } from '@angular/compiler/src/util';
 })
 export class FilterSearchPipe implements PipeTransform {
 	transform(items: any[], search: string) {
-		if (search) {
+		if (search && items) {
 			const words = search.split(',').filter(word => word).map(word => escapeRegExp(word));
 			const pattern = words.map((word, index) => (index < words.length - 1) ? word + '|' : word).join('');
 			const contains = new RegExp(pattern, 'i');
 			return items.filter(item => contains.test(item.path));
 		}
+
 		return items;
 	}
 }