+++ src/core/filter/filter.view.js
@@ -9,29 +9,29 @@ export class FilterView {
 			source: 'filter.view',
 			execute: (column, search) => {
 				const { key } = column;
-				const { filter } = model;
 
-				const by = clone(filter().by);
-				if (!isUndefined(search) && search !== '') {
-					by[key] = {
-						expression: {
-							kind: 'group',
-							op: 'and',
-							left: {
-								kind: 'condition',
-								left: key,
-								op: 'like',
-								right: search
-							},
-							right: null
-						}
-					};
+				let { by, operatorFactory } = model.filter();
+				by = clone(by);
+				const filter = by[key] || (by[key] = {});
+				if (!isUndefined(search) && search !== null && search !== '') {
+					const opList = operatorFactory(column);
+					const op = filter.expression ? filter.expression.op : opList[0];
+					if (op === 'contains') {
+						filter.items = [search];
+					} else {
+						filter.expression = {
+							kind: 'condition',
+							left: key,
+							op,
+							right: search
+						};
+					}
 				}
 				else {
 					delete by[key];
 				}
 
-				filter({ by });
+				model.filter({ by }, { source: 'filter.view' });
 			}
 		});
 	}
@@ -39,5 +39,20 @@ export class FilterView {
 	has(column) {
 		const { by } = this.model.filter();
 		return by.hasOwnProperty(column.key);
+	}
+
+	value(column) {
+		const { key } = column;
+		const { by } = this.model.filter();
+		if (by[key]) {
+			const { expression, items } = by[key];
+			return expression
+				? expression.right
+				: items && items.length
+					? items[0]
+					: null;
+		}
+
+		return null;
 	}
 }