+++ src/core/pipe/units/scene.pipe.unit.js
@@ -1,0 +1,13 @@
+import { Pipe } from '../pipe';
+
+export const scenePipeUnit = [
+	Pipe.data,
+	Pipe.memo,
+	Pipe.column,
+	Pipe.view,
+	Pipe.pagination,
+	Pipe.animation,
+	Pipe.scene
+];
+
+scenePipeUnit.why = 'redraw';