+++ src/core/pipe/units/view.pipe.unit.js
@@ -4,7 +4,8 @@ export const viewPipeUnit = [
 	Pipe.data,
 	Pipe.memo,
 	Pipe.column,
-	Pipe.view
+	Pipe.view,
+	Pipe.scene
 ];
 
 viewPipeUnit.why = 'refresh';