+++ src/core/selection/selection.view.js
@@ -87,7 +87,16 @@ export class SelectionView {
 
 		const toggleActiveRow = new Command({
 			source: 'selection.view',
-			canExecute: () => model.selection().unit === 'row' && this.rows.length > 0,
+			canExecute: () => {
+				const { rowIndex } = model.navigation();
+				const row = this.rows[rowIndex >= 0 ? rowIndex : rowIndex + 1];
+
+				if (!this.form.canSelect(row)) {
+					return false;
+				}
+
+				return model.selection().unit === 'row' && this.rows.length > 0;
+			},
 			execute: () => {
 				const { rowIndex } = model.navigation();
 				const row = this.rows[rowIndex >= 0 ? rowIndex : rowIndex + 1];
@@ -140,7 +149,7 @@ export class SelectionView {
 
 					const e = {
 						items: isUndefined(row)
-							? this.model.view().rows
+							? this.model.scene().rows
 							: [row],
 						source: 'custom',
 						kind: 'toggleRow'
@@ -287,7 +296,7 @@ export class SelectionView {
 		const { toggle } = this.model.selection();
 
 		items = !arguments.length || isUndefined(items)
-			? this.model.view().rows
+			? this.model.scene().rows
 			: isArray(items)
 				? items : [items];